split up lol code
