+++ lib/leagueoflegends/help.js
@@ -1,0 +1,19 @@
+export const help_text = `League of Legends help commands
+
+Accepted positions are **Top**, **Mid**, **ADC**, **Jungle**, and **Support**.
+Accepted regions are **br, eune, euw, kr, lan, las, na, oce, pbe, ru, and tr**.
+
+**\`!lol bans\`**
+    Get the top 10 most common bans
+**\`!lol best\`** \`position\`
+    Get the top 10 best champs for a position
+**\`!lol counters\`** \`champ-name\` \`position\`
+    Get the top 10 counters for a Champion and Position
+**\`!lol items\`** \`champ-name\` \`position\`
+    Get the highest win item sets for a Champion and Position
+**\`!lol match\`** \`region\` \`summoner-name\`
+    Get rank, champ, winrate, and games for all players in a __current__ match.
+**\`!lol skills\`** \`champ-name\` \`position\`
+    Get the highest win skills for a Champion and Position
+**\`!lol status\`**
+    Get the LoL Game and Client server status for all regions`;
